+++ cmem_plugin_pyshacl/plugin_pyshacl.py
@@ -452,10 +452,8 @@ class ShaclValidation(WorkflowPlugin):
                 replace=self.clear_validation_graph,
                 content_type="application/n-triples",
             )
-            if res.status_code != 204:  # noqa: PLR2004
-                raise OSError(
-                    f"Error posting SHACL validation graph (status code {res.status_code})."
-                )
+        if res.status_code != 204:  # noqa: PLR2004
+            raise OSError(f"Error posting SHACL validation graph (status code {res.status_code}).")
 
     def check_object(  # noqa: C901 PLR0912 PLR0913
         self, graph: Graph, subj: Node, pred: URIRef, data_graph: Graph, shacl_graph: Graph